+++ src/core/LocalState.ts
@@ -40,6 +40,7 @@ export type Resolver = (
   context?: any,
   info?: {
     field: FieldNode;
+    fragmentMap: FragmentMap;
   },
 ) => any;
 
@@ -390,7 +391,7 @@ export class LocalState<TCacheShape> {
             rootValue,
             argumentsObjectFromField(field, variables),
             execContext.context,
-            { field },
+            { field, fragmentMap: execContext.fragmentMap },
           ));
         }
       }